+++ omicverse/externel/__init__.py
@@ -1,4 +1,4 @@
 from . import (scSLAT,CEFCON,mofapy2,GNTD,spaceflow,STT,
                tosica,STAGATE_pyG,STAligner,spatrio,PROST,cytotrace2,
-               GraphST,commot,cnmf,starfysh,scMulan,flowsig,PyWGCNA
+               GraphST,commot,cnmf,starfysh,scMulan,flowsig,PyWGCNA,CAST
                )